게시판 댓글 댓글 작성자만 수정 및 삭제 로직 구현

+++ src/main/resources/static/js/app/index.js
@@ -151,7 +151,17 @@ var main = {
                 id: form.querySelector('#id').value,
                 postsId: form.querySelector('#postsId').value,
                 comment: form.querySelector('#comment-content').value,
+                writerUserId: form.querySelector('#writerUserId').value,
+                sessionUserId: form.querySelector('#sessionUserId').value,
             }
+            console.log("commentWriterID : " + data.writerUserId);
+            console.log("sessionUserID : " + data.sessionUserId);
+
+            if (data.writerUserId !== data.sessionUserId) {
+                alert("본인이 작성한 댓글만 수정 가능합니다.");
+                return false;
+            }
+
             if (!data.comment || data.comment.trim() === "") {
                 alert("공백 또는 입력하지 않은 부분이 있습니다.");
                 return false;
@@ -172,20 +182,24 @@ var main = {
             }
         },
         /** 댓글 삭제 */
-        commentDelete : function (postsId, commentId) {
+        commentDelete : function (postsId, commentId, commentWriterId, sessionUserId) {
 
+        if (commentWriterId !== sessionUserId) {
+            alert("본인이 작성한 댓글만 삭제 가능합니다.");
+        } else {
             const con_check = confirm("삭제하시겠습니까?");
-            if (con_check === true) {
-                $.ajax({
-                    type: 'DELETE',
-                    url: '/posts/' + postsId + '/comments/' + commentId,
-                    dataType: 'JSON',
-                }).done(function () {
-                    alert('댓글이 삭제되었습니다.');
-                    window.location.reload();
-                }).fail(function (error) {
-                    alert(JSON.stringify(error));
-                });
+                        if (con_check === true) {
+                            $.ajax({
+                                type: 'DELETE',
+                                url: '/posts/' + postsId + '/comments/' + commentId,
+                                dataType: 'JSON',
+                            }).done(function () {
+                                alert('댓글이 삭제되었습니다.');
+                                window.location.reload();
+                            }).fail(function (error) {
+                                alert(JSON.stringify(error));
+                            });
+                        }
             }
         }
 };
